Add support for data URLs with empty mimetypes, like `data:,abc`.
Per the data URL spec, the mimetype segment of a data URL should be
optional. The change here, which might be hard to read in a diff, is
adding an extra pair of parentheses and question mark to the regular
expression to make the mimetype optional.

I've also added a test that fails without the change and verifies that
we can construct and render a URL without a specified mimetype.